compile_charclass: rewrite to use danglin' root
This is the first step in making the charclass compiler more testable,
and towards creating a reverse NFA for charclasses.

Before, we stored a "dummy" [00-FF] node and referenced that as the root
node. Now, we treat the very first node inserted as the root node, and all functions have been
updated to match.

Additionally, the size of re__compile_charclass_tree was greatly reduced
by omitting the prev_sibling_ref and last_child_ref members. Now, we use
a pretty slick trick that when creating the tree, stores child_ref as if
it were last_child_ref, and storing sibling_ref as if it were
prev_sibling_ref. Then, when hashing and merging all of the nodes in the
tree, we reverse the sibling_ref so that it is next_sibling_ref, and
update child_ref so that it is first_child_ref. Building the tree in
reverse also allows us to avoid iterating through child nodes
whenever we want to add a new node, we just have to be able to
inspect the previous child for intersection, otherwise it's just a quick
append most of the time. re__compile_charclass_tree is nominally 16
bytes now, at least it is on my M1, and that's cool too. Should be
altogether a bit faster.