ast: add reverse iteration fields
Adds last_child_ref to the AST struct. This is another step towards
making a reverse program possible. The AST struct is now 32 bytes, but
it's not a huge deal since it was 28 bytes before (it may be even faster
now that it's aligned to some kind of special boundary or something)
Removed add_sibling, replacing its behavior in add_child. Since we have
last_child_ref inside all nodes, add_child is O(1) like add_sibling was.
Probably a good idea to unify the two.