exec_nfa: fork from exec
I split up the exec code between NFA and DFA. re__exec_nfa is the NFA
execution context, and re__exec_dfa is the DFA execution context.
Common code is still in re__exec_nfa.c, but I'll probably change this
later.
Also added are the beginnings of the DFA.