build: the vanity update
Moves to a new system of handling bits. In addition, many symbols are
now prefixed with my initials - mn. This isn't because I'm that
egotistical, it's because I plan to use these bits for many other
libraries, and want to be wary of symbol duplication, as that's
something people are bound to have a huge gripe with. I'm going to add
an option that lets users prefix all symbols as 'mn_re' instead of 'mn'
or just 're' if they desire.