dfa: add initial implementation
a case study in premature optimization if I've ever seen one. This DFA
uses a pretty standard construction, empty-width flags are embedded in
the state (as per burntsushi's HN comment). Notably doesn't make use of
an arena like many other instances in this project. to protect against
*bad* memory fragmentation I use a page system for storing thread sets
and a similar page system for storing the actual states. Thread sets are
chunked in blocks of 256 (at the moment) and then contiguous blocks of
threads are allocated from those larger blocks. For DFA states, I
allocate enough space for 4 at a time and then use that till it fills
up, at which point I allocate space for another 4. This is untested,
it might not be worth it.

For looking up states I use a rudimentary hash table with quadratic
probing and a load factor of 0.75. Ideally cache hits should be rare
but if they turn out to be more common this should be quite fast.

Right now only dfa_start() is implemented but dfa_run() should be easy
to do, I have the tools for it all set up in this commit. but for now
I'm gonna go over to Kenny's and eat some well-deserved grub.