nfa: big refactor
This commit refactors the NFA, and, to an extent, the DFA.
The original implementation of the NFA is incompatible with the DFA.
The NFA used to work by running an epsilon (nfa_start(assert)), and then
running combos of byte/epsilon (nfa_run(ch, assert)).
This doesn't work with the DFA, at least not easily. The DFA works by
running combos of byte/epsilon (dfa_run(ch)), and then a single epsilon
(dfa_run(END)). The NFA now mirrors this implementation.