charclass: fix edge case in builder_finish()
In some situations when making an inverted charclass, there is no need
to add a final range going from (x, RUNE_MAX) because the charclass
already encompasses this.
We already check for this actually, but it wasn't thorough enough.
We now check if the temporary maximum exceeds RUNE_MAX too.